+++ planner-app/src/main/java/dev/watchnest/plannerapp/persistence/jpa/JpaPersistenceConfig.java
@@ -1,0 +1,13 @@
+package dev.watchnest.plannerapp.persistence.jpa;
+
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@Configuration
+@Profile("persistent")
+@EnableJpaRepositories(basePackages = "dev.watchnest.plannerapp.persistence.jpa")
+@EntityScan(basePackages = "dev.watchnest.plannerapp.persistence.jpa")
+public class JpaPersistenceConfig {
+}